Unit_testing/Test_py.py
import unittest
import numpy as np
import kMeans
import phq
import dithering_module
import stitch_pattern_maker
from PIL import Image

class TestKMeans(unittest.TestCase):

    def test_kmeans_quantization(self):
        image = np.random.randint(0, 256, (100, 100, 3), dtype=np.uint8)
        quantized_image, centers = kMeans.kmeans_quantization(image, 4)
        self.assertEqual(quantized_image.shape, image.shape)
        self.assertEqual(centers.shape[1], 3)

    def test_create_pattern(self):
        image = np.random.randint(0, 256, (100, 100, 3), dtype=np.uint8)
        pattern = kMeans.create_pattern(image, 10)
        self.assertEqual(pattern.shape, (10, 10, 3))

class TestPHQ(unittest.TestCase):

    def test_kmeans_quantization(self):
        image = np.random.randint(0, 256, (100, 100, 3), dtype=np.uint8)
        histogram_r = np.histogram(image[..., 0], bins=256)[0]
        histogram_g = np.histogram(image[..., 1], bins=256)[0]
        histogram_b = np.histogram(image[..., 2], bins=256)[0]
        quantized_image, centers = phq.kmeans_quantization(image, histogram_r, histogram_g, histogram_b, 4)
        self.assertEqual(quantized_image.shape, image.shape)
        # Adjust the expected size of centers based on the function's implementation
        self.assertEqual(centers.shape[1], 6)

class TestStitchPatternMaker(unittest.TestCase):

    def test_stitch_pattern(self):
        image = np.random.randint(0, 256, (100, 100, 3), dtype=np.uint8)
        stitch_size = 10
        stitch_width = 5
        stitch_pattern = stitch_pattern_maker.stitch_pattern(Image.fromarray(image), stitch_size, stitch_width)
        # Adjust the expected size of the stitch pattern based on the function's implementation
        self.assertEqual(stitch_pattern.size, (100, 100))


class TestDitheringModule(unittest.TestCase):

    def test_floyd_steinberg_dithering(self):
        image = np.random.randint(0, 256, (100, 100, 3), dtype=np.uint8)
        dithered_image = dithering_module.floyd_steinberg_dithering(image)
        self.assertEqual(dithered_image.shape, image.shape)

if __name__ == '__main__':
    unittest.main()
