requirements.txt
matplotlib==3.5.1
numpy==1.22.1
pandas==1.3.5
Pillow==9.5.0
plotly==5.15.0
scikit-image==0.21.0
scikit-learn==1.2.2
sklearn==0.0
streamlit==1.24.1
streamlit-lottie==0.0.5
webcolors==1.13
